# developer-docs/api-reference/openapi.yaml
openapi: 3.1.0
info:
  title: GlobalTalk AI & Desi Language AI API
  description: |
    Official API specification for GlobalTalk AI and the Desi Language AI engine.
    Provides real-time neural translation, document translation, writing assistance,
    and native Indic linguistic intelligence across 100+ global languages and all 22 official
    Eighth Schedule Indian languages.
  version: 3.2.0
  contact:
    name: GlobalTalk AI Developer Relations
    url: https://globaltalk.ai/support
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.globaltalk.ai
    description: Global Production Edge Gateway
  - url: http://127.0.0.1:8088
    description: Local Development / Edge Instance
security:
  - ApiKeyAuth: []
  - DesiKeyAuth: []
  - BearerAuth: []
paths:
...
  /v2/document:
    post:
      summary: Upload Document for Translation
      description: |
        Asynchronously submits a document (PDF, DOCX, PPTX, XLSX, HTML, TXT) for layout-preserving translation.
      operationId: uploadDocument
      tags:
        - Document Translation
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/DocumentUploadRequest'
      responses:
        '200':
          description: Document successfully queued
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentUploadResponse'
  /v2/document/{document_id}:
    get:
      summary: Check Document Status
      description: Polls the processing status, progress percentage, and billing metrics of a document job.
      operationId: getDocumentStatus
      tags:
        - Document Translation
      parameters:
        - name: document_id
          in: path
          required: true
          schema:
            type: string
        - name: document_key
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Current document status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentStatusResponse'
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: GlobalTalk AI API Key (`gtk_...`)
    DesiKeyAuth:
      type: apiKey
      in: header
      name: X-Desi-Key
      description: Desi Platform API Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    DocumentUploadRequest:
      type: object
      required:
        - file
        - target_lang
      properties:
        file:
          type: string
          format: binary
        target_lang:
          type: string
          example: "ES"
        source_lang:
          type: string
        formality:
          type: string
          enum: [default, more, less]
        glossary_id:
          type: string
    DocumentUploadResponse:
      type: object
      properties:
        document_id:
          type: string
          example: "doc_9a7b8c12"
        document_key:
          type: string
          example: "key_sec_4f5e6d"
    DocumentStatusResponse:
      type: object
      properties:
        document_id:
          type: string
        status:
          type: string
          enum: [queued, translating, done, error]
          example: "done"
        seconds_remaining:
          type: integer
          example: 0
        billed_characters:
          type: integer
          example: 1420
...